refactor: remove middleware manager
